+++ deepinv/physics/singlepixel.py
@@ -138,30 +138,6 @@ def sequency_mask(img_size: tuple[int], m: int) -> torch.Tensor:
     return mask
 
 
-def old_sequency_mask(img_size: tuple[int], m: int) -> torch.Tensor:
-    """
-    Generates a binary mask for a single-pixel camera based on a sequency ordering.
-    :param tuple img_size: The shape of the input image as a tuple (C, H, W), where C is the number of channels,
-                            H is the height, and W is the width.
-    :param int m: The number of pixels to include in the mask, selected based on the sequency ordering.
-    :return: A binary mask of shape (1, C, H, W) with `m` pixels set to 1 and the rest set to 0.
-    :rtype: torch.Tensor
-    """
-
-    _, H, W = img_size
-    n = H * W
-
-    indexes = get_permutation_list(n)[:m]
-    i, j = np.meshgrid(np.arange(H), np.arange(W), indexing="ij")
-    i = i.flatten(order="F")
-    j = j.flatten(order="F")
-
-    mask = torch.zeros((1, *img_size))
-    mask[:, :, i[indexes], j[indexes]] = 1.0
-
-    return mask
-
-
 def cake_cutting_seq(i: int, p: int) -> list:
     """
     Generates a sequence based on the given index `i` and parameter `p`. The sequence alternates
@@ -335,14 +311,10 @@ class SinglePixelCamera(DecomposablePhysics):
     An existing operator can be loaded from a saved ``.pth`` file via ``self.load_state_dict(save_path)``,
     in a similar fashion to :class:`torch.nn.Module`.
 
-    .. warning::
-
-        Since version 0.3.1, a small bug in the sequency ordering has been fixed. However, it is possible to use the old sequency ordering by setting `ordering='old_sequency'`.
-
     :param int m: number of single pixel measurements per acquisition (m).
     :param tuple img_size: shape (C, H, W) of images.
     :param bool fast: The operator is iid binary if false, otherwise A is a 2D subsampled hadamard transform.
-    :param str ordering: The ordering of selecting the first m measurements, available options are: `'sequency'`, `'cake_cutting'`, `'zig_zag'`, `'xy'`, `'old_sequency'`.
+    :param str ordering: The ordering of selecting the first m measurements, available options are: `'sequency'`, `'cake_cutting'`, `'zig_zag'`, `'xy'`.
     :param torch.Generator rng: (optional) a pseudorandom random number generator for the parameter generation.
         If ``None``, the default Generator of PyTorch will be used.
 
@@ -408,12 +380,6 @@ class SinglePixelCamera(DecomposablePhysics):
                 mask = zig_zag_mask(img_size, m)
             elif ordering == "xy":
                 mask = xy_mask(img_size, m)
-            elif ordering == "old_sequency":
-                # Raise warning if the old sequency mask is used
-                print(
-                    "Warning: The old sequency mask is deprecated. Plase, use sequency mask instead."
-                )
-                mask = old_sequency_mask(img_size, m)
             else:
                 raise ValueError(
                     f"Unknown ordering {ordering}. Available options are: `sequency`, `cake_cutting`, `zig_zag`, `xy`."